Muck around with Python in a pretty-much-not-helpful way
Was thinking about the idea that configuration languages should be
Turing-complete, because that's the only way to truly guarantee that
redundancy can be eliminated from your configuration. Since YAML is a
superset of JSON, I wrote a Python module that emits Docker Compose
YAML files by using Python's JSON support (therefore no external
dependency is needed, and since the files don't have to be really
human-readable there's no cost to not using a real YAML library).

By doing this I was able to reduce the 32-line simulator.yml and the
24-line rviz.yml with a 10-line simulator.py and a 1-line rviz.py,
with some loss of legibility but I don't think too much; I've also
added the ability to specify custom packages to install in the (now
Python instead of YAML) compose definitions directly, which means we
could remove rviz from the list of all packages installed from
ros/Dockerfile and instead install it only in the compose file that
needs it directly. This allows an architecture wherein the `ros`
directory builds a container that only has the set of packages that
contain message definitions, and we install telemetry tools like rviz
only in containers that need them.

If we make use of *only* compose files for running major Docker
actions, we can avoid having to write complicated docker invocations
in different scripts or relying on some images already having been
built under particular names; and then we can use just the one
invocation of `docker compose` in `containers.py` to avoid things like
forgetting to write `--build` in the parameter list, etc. I've also
made `containers.py` automatically run `docker compose down` after the
container finishes (whether it succeeded or not), meaning we don't
have any leftover container networks or anything lying around after
we're done; and this means that my `rviz` container implementation
fails immediately with a "missing network" error as soon as the
container starts, rather than starting successfully and trying to
connect to a ROS master that's not there. We also get the ability to
do the `xhost` thing without having to explicitly write it out, and we
get specify in `roscmd()` where the ROS master URL is.